social product fixes
