+++ src/client/social-gifts/SocialProduct/PicturesSlider/PicturesSlider.less
@@ -8,6 +8,7 @@
     object-fit: contain;
     width: 350px;
     height: 350px;
+    cursor: pointer;
     @media @small {
       width: 500px;
       height: 500px;
@@ -17,7 +18,7 @@
     object-fit: contain;
     width: 50px;
     height: 50px;
-    border: 1px solid @normal-grey;
+    border: 1px solid @silver-gray;
     border-radius: 4px;
     &:hover,
     &-active {
@@ -45,7 +46,7 @@
       font-size: 20px !important;
       display: inline-block;
       z-index: 1200;
-      color: #bebebe !important;
+      color: @silver-gray !important;
     }
   }
 }
