+++ src/client/social-gifts/SocialProduct/SocialProduct.less
@@ -364,6 +364,7 @@
     background-color: var(--website-light-color);
     border: 1px solid var(--website-color);
     margin: 0 0 10px 10px;
+    color: @black;
   }
   .ant-carousel .slick-dots li button {
     background: @normal-grey;
